refactor: consolidate Jest configuration
- Moved Jest configuration to a new base file for better reusability.
- Updated backend and frontend Jest config imports to reference the new base configuration.
- Removed the old Jest configuration file from the apps directory.
