--- apps/jest.config.base.js
@@ -1,7 +1,0 @@
-export default {
-  moduleFileExtensions: ['js', 'json', 'ts'],
-  transform: {
-    '^.+\\.(t|j)s$': 'ts-jest',
-  },
-  collectCoverageFrom: ['**/*.(t|j)s', '!**/*.spec.ts', '!**/node_modules/**'],
-};
